Add files via upload
This repository contains the code for a nudity prediction project. Please note that the dataset containing explicit content is not included in this repository due to its sensitive nature. However, you can find the link to access the dataset in the following section.

# Dataset

The dataset used for this project consists of explicit content images. To access the dataset, please [click here](https://drive.google.com/file/d/1jBqWt9YtsROOTnsB4QMPJQaK8JNzxCJd/view?usp=share_link).

# Disclaimer

The explicit content dataset provided is intended for research purposes only. By accessing and using the dataset, you assume all responsibility and liability. We strongly advise using the dataset in compliance with applicable laws, regulations, and ethical guidelines. The repository owner and contributors are not responsible for any misuse or legal issues arising from the usage of this dataset.

# Usage

To utilize the code in this repository, follow the instructions below:

Clone the repository to your local machine.
Install the necessary dependencies as outlined in the requirements file.
Run the scripts for data preprocessing, model training, and testing as per your requirements.
Customize and modify the code as needed for your specific use case.
Contributing

Contributions to this project are welcome. If you have any suggestions, bug reports, or feature requests, please open an issue or submit a pull request.

License

The code in this repository is licensed under the MIT License. However, please note that this license does not cover the explicit content dataset. Refer to the dataset source for its specific terms of use and licensing information.

Please use the provided dataset responsibly and ensure compliance with legal and ethical guidelines.